+++ src/main/java/com/prestashop/config/BaseUrl.java
@@ -1,0 +1,5 @@
+package com.prestashop.config;
+
+public class BaseUrl {
+    public static final String PRESTASHOP = "https://demo.prestashop.com/#/en/front";
+}
