+++ src/main/java/com/prestashop/webdrivers/Browser.java
@@ -1,0 +1,12 @@
+package com.prestashop.webdrivers;
+
+import org.openqa.selenium.WebDriver;
+
+public interface Browser {
+    /**
+     * Gets a Web Driver instance of a specific browser.
+     *
+     * @return Web Driver instance.
+     */
+    WebDriver getDriver();
+}
