+++ src/main/java/com/prestashop/webdrivers/DriverType.java
@@ -1,0 +1,7 @@
+package com.prestashop.webdrivers;
+
+public enum DriverType {
+    CHROME,
+    FIREFOX,
+    REMOTE
+}
