--- src/test/java/config/BaseUrl.java
@@ -1,5 +1,0 @@
-package config;
-
-public class BaseUrl {
-    public static final String PRESTASHOP = "https://demo.prestashop.com/#/en/front";
-}
