--- src/test/java/stepdefinitions/sharedSteps/SharedSteps.java
@@ -1,12 +1,0 @@
-package stepdefinitions.sharedSteps;
-
-import io.cucumber.java.en.When;
-import utils.DriverActions;
-
-public class SharedSteps {
-
-    @When("user open previous page in browser")
-    public void userOpenPreviousPageInBrowser() {
-        DriverActions.switchToThePreviousPageInBrowser();
-    }
-}
