--- src/test/java/webdrivers/Browser.java
@@ -1,12 +1,0 @@
-package webdrivers;
-
-import org.openqa.selenium.WebDriver;
-
-public interface Browser {
-    /**
-     * Gets a Web Driver instance of a specific browser.
-     *
-     * @return Web Driver instance.
-     */
-    WebDriver getDriver();
-}
