--- src/test/java/webdrivers/DriverType.java
@@ -1,7 +1,0 @@
-package webdrivers;
-
-public enum DriverType {
-    CHROME,
-    FIREFOX,
-    REMOTE
-}
